requirements.txt
aioschedule==0.5.2
bs4==0.0.1
pymongo==3.7.1
python-decouple==3.4
python-dotenv==0.20.0
python-telegram-bot==20.3
requests==2.27.1
sentry-sdk==1.5.12
beautifulsoup4==4.12.2
redis==5.0.1